Absorb EinsplineSetBuilderReadBands_ESHDF.cpp into EinsplineSetBuilderESHDF.fft.cpp
